requirements.txt
pdfminer.six
transformers
nltk
chromadb
langchain
langchain-openai
langchain-chroma
langchain-community
dashscope
pypdf
